move network capabilities to common
